Modularized Testing fixtures
