+++ schedule/schedule-domain-test/src/main/java/net/squanchy/schedule/domain/view/ScheduleFixtures.kt
