+++ speaker/speaker-domain-test/.gitignore
@@ -1,0 +1,1 @@
+/build
