+++ speaker/speaker-domain-test/src/main/java/net/squanchy/speaker/domain/view/SpeakerFixtures.kt
